test that contract sets total supply to one million
import contract using require statement then instantiate and deploy contract. because totalSupply is a state variable in the contract it comes with a getter function. Because javascript treats this as a big number (<BN>), we format with toNumber() method and, utilizing mocha/chai library, assert it to be equal to what we assigned to it in the contract (one million in this case).