token total supply now defined upon contract instantiation
Modifies constructor function in dappToken.sol contract to accommodate an uint256 _initialSupply parameter which is used to set the totalSupply upon instantiation rather than hard coded into contract.
Beginning local variable names with an underscore is a convention used in Solidity to distinguish them from state variables. In this case _initialSupply is within the scope of the constructor function while totalSupply is a state variable available to the entire contract.
Also modifies the 2_deploy_contracts.js to instantiate the contract with a totalSupply of one million. This is accomplished via the deploy() function which can accept arguments to pass to the constructor function.